jit: irlower-call: Do proper loading in cgCheckRefs to eliminate truncation
The vasm for CheckRefs distinguishes between 8-bit, 32-bit and 64-bit
masks. However it uses a 64-bit load for all cases in combination with
a movtqX. This patch changes this to use a load which does not require
a truncation for the cost of being little-endian specific (which is a
requirement for the cgCheckRefs() anyways).

Testing the all set showed no regressions on ARM64.

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>