get stacktraces from initialize failure
Summary:
I was getting exceptions raised by ClientIde.connect_persistent but we'd lost the details of the stack trace that caused them. This restored it.

(This will be superseded by Waleed's forthcoming diff. I'm putting this out so we can remember to make sure the same trace gets captured by Waleed.)

Reviewed By: arxanas

Differential Revision: D4850996

fbshipit-source-id: a3055c33db923304750501f5ea7b2eddfad0cfe4