Reject equality operator on Awaitable
Summary:
Currently we permit the use of `===` and `!==` on values of type `Awaitable<_>` even though this is very likely an error. The programmer probably mean to write an
`await` statement and assign to a local first. So instead, we now reject any use of equality on values that have `Awaitable` type.

Reviewed By: hubyrod, oulgen

Differential Revision: D4834654

fbshipit-source-id: 09a842c4c34fd849410cdbbd6ec2bc017b5bd2db