Avoid calling Unit::lookupFunc from the JIT
Summary:
In a few places we call Unit::lookupFunc to map a name to a Func pointer. The
problem with this is Unit::lookupFunc uses RDS, and we don't want the JIT
reading from RDS. Its also dodgy because it mixes up the runtime state with the
compiler's state. In fact, now that the normal RDS section is swapped out at JIT
time, it means we'll never lookup a non-persistent function in Unit::lookupFunc.

Instead, provide a new function called lookupImmutableFunc. This function
provides a function which always uniquely maps to a name in a given
context. Note that these functions aren't always persistent. Use this function
to lookup functions in the JIT.

No functions are immutable if JitRenameEnableFunction is enabled. Nonetheless,
sometimes we want to lookup a builtin inside the compiler without necessarily
caring if its immutable. Add Unit::lookupBuiltin for this purpose.

In order to facilitate all this, store a separate Func* pointer in NamedEntity,
which represents the unique function which maps to that name (if any).

Reviewed By: markw65

Differential Revision: D4855614

fbshipit-source-id: 03d1c004823e2fae5b34b3720c261a343494a66e