Fix SkipFrame handling for xdebug
Summary:
Some of the xdebug code wasn't taking into SkipFrame into account
properly. It could attach a VarEnv to a builtin (which is skip-frame), then when
it tries to eval, it would skip over the builtin while looking for the previous
frame. This triggers an assert because there shouldn't be any VarEnvs between
the new frame and the old frame.

After this was fixed, there was a second issue where we rewrite vmfp() before
doing the eval, but not vmpc(). This can trigger asserts because we verify that
vmpc() lies within vmfp()'s code in a few places. Fix this by rewriting vmpc()
as well.

Reviewed By: mxw

Differential Revision: D4837401

fbshipit-source-id: 6c10ca7494489c53ec3f211153e9e23c75a5413f