Apply expression dependent types to unresolved types before losing the
Summary:
After some debugging, I realized that if we have a type access on a type such
as

<static> as this as TUnresolved(Foo)

and we try to access type T on it, when we run typing_taccess, we don't
preserve the env.dep_tys from each elemeent of the TUnresolved's type list.
Thus we lose type access information. For example, in the above example, Foo
may expaned to Foo::T, but <static> and this don't get the "::T" added to them
because we throw away the environment.

This diff changes that and fixes the behavior by applying expression dependent
types earlier.

I'm not sure how we avoided this behavior before because it clearly looks broken to me, but I expect it's because we reapplyed the expression dependent types during subtyping instead of storing the broken ones.

Differential Revision: D4796482

fbshipit-source-id: d7b20015341571e9edf7761d423e90ea84379c8e