Fix crash related to lazy rules and async blocks
Summary:
async blocks can be used as the left-hand side of the pipe operator.

The problem was we were ending the chunk group at the end of the async block even though the chunk group should also contain the right-hand side of the pipe operator. Since binary operators use lazy rules, and lazy rules don't persist across chunk groups, the code was trying to look up a rule that was in the previous chunk group, leading to fatals.

This code fixes that problem

Differential Revision: D4856571

fbshipit-source-id: bf7589736ecabc8b655655b9b430ce8dd43d90bd