Fix an issue with FPass* and hack arrays
Summary:
dict_set, keyset_set and vec_set would all convert TInitGen to
TInitCell before calling array_like_set. This meant they could return
a noThrow flag, even when they were certain to throw (and eg
miFinalVGetElem was about to call unreachable(env)).

If in addition the element was known to exist in the array, this could
cause an assert in mergePaths because both the cget and the vget side
of the FPass would claim to be nothrow.

Reviewed By: ricklavoie

Differential Revision: D4860185

fbshipit-source-id: 8e27aebbaa099033c65a10eb04e680f54800f67b