Fix static_assert in memory-manager.h
Summary:
The assert was comparing kNumSmallSizes to sizeof(kSmallSize2Index), but they
are not related; in fact, sizeof(kSmallSize2Index) was arbitrarily chosen to
be 256 (tradeoff between faster lookups and table size). Compare kNumSmallSizes
to the number of entries in kSmallIndex2Size instead; kSmallIndex2Size has an
entry for every size class, and kNumSmallSizes must be <= the number of size classes.

Reviewed By: edwinsmith

Differential Revision: D4875225

fbshipit-source-id: f3e137c42b9475dd7a083a5e79ecc45e276242a9