Redis.php: remove explicit QUIT in close function
Summary:
The Redis QUIT always return "OK", but in Redis.php close()
function the TCP socket is closed immediately after sending
the command. This causes frequent TCP RST returned by HHVM
to the Redis server when the latter sends the "OK" response.

To fix the issue the QUIT command is removed from the close()
function and added as function ('quit', via __call). This allows
users to "quit() and close()" if needed.

The same issue has been discussed/fixed in phpredis:
https://github.com/phpredis/phpredis/issues/562
Closes https://github.com/facebook/hhvm/pull/7771

Differential Revision: D4851189

Pulled By: mofarrell

fbshipit-source-id: 251ae15bc9ab91553d9b64ad6331cac727cf2305