Indent true/false expressions in ConditionalExpression
Summary:
Increase the indentation level of the expressions in the branches of a ConditionalExpression, so that they start and end at the same indentation level (due to the coincidence of "? " and ": " being the same width as our indentation level).

Since this doesn't look great at other indentation widths, don't indent unless the indent width is 2. This is currently always true, but we will likely make it configurable in the future.

Since prettier also indents these subexpressions, this change will make the formatting of our JS and Hack more consistent.

Reviewed By: eshrews

Differential Revision: D4869647

fbshipit-source-id: fc05d19e7283db5909948c1ab04f8446edf227c0