Loosen typing strictness for Shapes::idx
Summary: Adjusted the typing rule for `Shapes::idx` to allow accessing a field of a partial shape that does not define or unset the field. Added a lint warning that the shape may be instantiated with the field, but with the wrong type -- this will conflict with the assumed return type of `Shapes::idx` and will not be caught by the typechecker. This change (and lint warnings) are guarded by experimental feature `experimental_shape_idx_relaxed`.

Reviewed By: michaeltingley

Differential Revision: D4835699

fbshipit-source-id: fbd3bdfeab13ee945ba7ef78ca46b893040be42b