Handle some unusual Bytecodes properly
Summary:
There are a handful of Base* bytecodes that read the stack, but
without popping the element read. We should handle them properly.

Currently, they're not a problem, because the final op always pops the
cell from the stack - so *it* marks it as used, and we don't allow
arbitrary instructions in minstr sequences. But for correctness (and
in case we ever start emitting SetMs that leave the base on the stack)
we should explicitly use the value.

Reviewed By: paulbiss

Differential Revision: D4817793

fbshipit-source-id: e6210d63ded8565c7f41b34a0911388f1d7a18da