tests: Add patch for openssl.cnf import.
This commit adds a patch for openssl.cnf, when obtaining the tests
via import_zend_test.py. The patch adds the key-value-pair
'default_md = sha256' to this file in order to circumvent
a security restriction on RedHat-based Linux distributions,
where OpenSSL does not validate MD5 certificates (which is
the default if default_md is not set).

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>