Make splits around single arguments not care about child splits
Summary:
If there is only one element in an argish list, and it had no surrounding whitespace in the original source, then don't split around it just because the element has child splits.

This preserves the pattern of putting the opener for a multi-line argument on the same line as the function name, to improve readability and decrease rightward drift. This pattern is so common that it surprises me that we haven't had a complaint about it yet. The test case in xhp.php is an example of code that was formatted particularly badly before this change.

Reviewed By: eshrews

Differential Revision: D4885588

fbshipit-source-id: ed4c17f316f6a333e52e0e5c9b41ead8d2179d72