Fix duplicated lines in XHP range formatting
Summary:
When formatting a range that ended between a BracedExpression and an XHPClose, the XHPClose line would be duplicated in the output.

This was due to some errors in our trivia handling, which sort of worked previously by coincidence. I think we're now correctly keeping track of which nodes actually do not have trailing trivia scanned by the lexer.

Reviewed By: eshrews

Differential Revision: D4887798

fbshipit-source-id: 6af955010ef3e674a9e7075550b7a7223af65d45