jit: irlower-call: Do proper loading in cgCheckRefs to eliminate truncation
Summary:
The vasm for CheckRefs distinguishes between 8-bit, 32-bit and 64-bit
masks. However it uses a 64-bit load for all cases in combination with
a movtqX. This patch changes this to use a load which does not require
a truncation for the cost of being little-endian specific (which is already
a requirement for the cgCheckRefs() function).

Testing the all set showed no regressions on ARM64.

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>
Closes https://github.com/facebook/hhvm/pull/7772

Differential Revision: D4868645

Pulled By: mxw

fbshipit-source-id: 7d515f1a15127c0572811c2130def26a03e1bd3b