Fix some repo-global issues
Summary:
While investigating the dce crash, I found that my new test case
crashed when I ran hhbbc as a separate process, but was fine when run
as part of hphpc.

After some digging, it was due to DisallowDynamicVarEnvFuncs. When
invoked with default arguments, hphpc left it set to false. With
suitable config options it would set it to true, but in either case
hphpc failed to write it to the global data. So when run as part of
hphpc, hhbbc respected the option - but since the option wasn't
written to global data, it wasn't necessarily respected at runtime.

When hhbbc was run standalone, it didn't read the option from global
data, defaulted it to true, and always wrote it to global data.

This fixes that so its always passed through global data; I also
noticed that hphpc writes APCProfile to the global data, but hhbbc
neither reads nor writes it; so it gets lost if we run hhbbc
separately.

Having fixed that, several tests failed because their options were
setup incorrectly; and after fixing that and the expect files one test
failed because the resulting regex was too complex for preg_match.
I fixed that by also checking the line-by-line diff. This is safe because
although the line-by-line diff can give false positives, it can't give
false negatives.

Reviewed By: ricklavoie

Differential Revision: D4904285

fbshipit-source-id: 17a32c3c4cc0bbbd12969a6db3ea137a0df0747c