Better type resolution
Summary:
We can do a better job of resolving the context class than looking it
up by name. This won't do much in practice, because we don't have
non-unique classes in our code base; but its cleaner, and more efficient.

We didn't resolve self and parent at all for type constraints, so fix
that, and also improve the handling of special interfaces.

Reviewed By: ricklavoie

Differential Revision: D4911712

fbshipit-source-id: d43adfbb12631adf0eaf9d28d14624672abce235