jit: code-gen-helpers: add byte truncation to emitCmpTVType().
The aarch64 code had an explicit byte-truncation
on the input register before performing a cmpbi operation.
When removing that HHVM triggers a segmentation fault.
The reason is, that cmpbi does not match anymore.
This is a bug on the vasm generation side, since
movtqb{} requires the register argument to be
of type Vreg8 and Vreg8 objects are always truncated
in the arm backend.

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>