jit: irlower: Mask type with kDataTypeMask before comparing.
When using 64-bit registers in cmpbi{} vasm instructions,
a preceding truncation to 8-bits is required in general.
This rule seems to be violated in some parts of the irlower
code. With the help of the arm64 backend, these locations
could be detected.

This patch uses masking with KindOfRefCountThreshold (0x7f)
to address this issue.

No regressions with the 'all' test set on arm64.

Note, that this patch is required in order to allow
a performance improvement patch for arm64 (drop emitting
uxtb instructions before cmpbi{} and cmpb{}).

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>