jit: vasm-lower: Add ARM specific truncation for type values.
When using 64-bit registers in cmpbi{} vasm instructions,
a preceding truncation to 8-bits is required in general.
This rule seems to be violated in some parts of the irlower
code. With the help of the arm64 backend, these locations
could be detected.

Note, that this patch is required in order to allow
a performance improvement patch for arm64 (drop emitting
uxtb instructions before cmpbi{} and cmpb{}).

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>