Implement goto label and statement naming
Summary:
Introduces naming and related checks for the goto feature. There are two primary enforcements resulting from this diff:

  1. A label may be declared at most once within a given function.
  2. A label may be referenced within a given function only if it is also declared within that function.

This diff includes many individual test cases for behavior that we may want to change in the future. These test cases serve as a form of documentation as to how goto statements are allowed to behave. I might have to take a second look here if there are limitations with how HHVM implements the goto functionality, but we can re-address that in the future.

Reviewed By: ericlippert

Differential Revision: D4826381

fbshipit-source-id: 1b669736ab7eee6b38654587bb32d6dbe98aed93