Fix test location and use test runner flags
Summary:
The lazy_rule_with_trivia test was not being run with the --test flag because it was not in the same directory as our HH_FLAGS file, causing sandcastle machines to log to Scuba.

This change moves the lazy_rule_with_trivia test to the right directory and switches to setting flags in the test runner instead of in directories, since we don't currently have any directory-specific needs.

Reviewed By: eshrews

Differential Revision: D4899263

fbshipit-source-id: 26a1e2ccedf703d268147bc8436d37caf3d9c66c