Definitions in statement positions for PHP support
Summary:
PHP allows for declarations in statement positions. A concrete example from the HHVM test set (`FPassC-2.php`):
```
if (1) {
  function g(&$x) {
    var_dump($x);
    $x = 123;
  }
}
```

If we want to support code generation for PHP, this type of "design" needs to be supported, so we need a representation for it in the `Ast` types. This diff suggests one way to do it.

Reviewed By: hubyrod

Differential Revision: D4862058

fbshipit-source-id: 4ddcd988cf3161798a99743e002799365a08d5dd