Add refining of load type-parameters in load-elim
Summary:
Even if load-elim cannot make a load redundant (replace it with an existing
SSATmp), it may still be able to prove the type-parameter on the load isn't as
refined as it could be. In that case, adjust the type-parameter to match the
more refined type. This is helpful, for example, to recover type information
lost by frame-state because of loop headers.

Reviewed By: markw65

Differential Revision: D4886883

fbshipit-source-id: dbb7bd3f5c6624ea64b4e8b06b6c7cf67d435834