Improve the coroutine closure's generated classname
Summary:
Fixes two issues that existed previously:

  1. We want the coroutine's enclosing class's name to be part of the generated state machine's name, since multiple classes may have coroutines with the same name. I've also replaced the term `StateMachine` with `Closure`.
  2. Previously, I was simply taking the `text` of the `function_name` node. This is not good, as it includes trivia. Instead, I now `syntax` the `function_name` node in order to get the underlying `Token`. From the `Token`, it is easy to get the name without trivia. This also applies to the classname.

Fixing #2 correctly required a bit of reworking. I wanted to avoid re-`syntax`ing the classname multiple times (and it should also be done as early as possible to short circuit in the case of unexpected input). Consequently, I pulled it all the way up into when we first `syntax` the `ClassishDeclaration`, and then threaded it down through everything else. I used a similar approach for the `function_name`, and factored out a little bit of duplicated logic from `coroutine_method_lowerer.ml` and `coroutine_state_machine_generator.ml` into `coroutine_lowerer.ml`.

This "threading through" could get unwieldy in the future if we need to thread more things through, so I'm very open to suggestions on how to improve this. IMO, it's still pretty manageable, so it's probably OK to punt for now.

Reviewed By: ericlippert

Differential Revision: D4931102

fbshipit-source-id: dae7b45c80a7dce5e577b8ffdb86448848f019a8