Fix an incorrect assert
Summary:
I had assumed that enums had to be defined as int or string, because
the runtime fatals if they are not - but there's nothing to stop them
being defined as another enum, or a typealias that eventually resolves
to an int or string.

This restructures the code so that we chase enum TypeConstraints in
the same loop as TypeAliases (with checks for circular definitions).

Reviewed By: ricklavoie

Differential Revision: D4943852

fbshipit-source-id: 318050e1984a76835f0200e0a1d552d4b1caada4