Adjustments to live rechecking behavior, part n-th
Summary:
As evidenced by [hate mail](https://fb.facebook.com/groups/hackforhiphop/permalink/1411736635541614) that I'm constantly getting, this is still not working as expected. The main problem is that at the moment, Hack server has three sourced of input:

- Nuclide
- Watchman
- Command line client

They race, and race conditions are hard. I went through a bunch of possible combinations and tested them, which revealed several bugs, but the truth is that no matter what strategy we choose to order them, there is always a scenario where it ends up with not intuitive behavior. Besides fixing bugs, this diff adjusts this ordering so that (I hope) the bad ordering is less likely too happen (see the test plan for examples).

Reviewed By: arxanas

Differential Revision: D4945655

fbshipit-source-id: 7d631ba02f6569011c783cea8de2d1a085f2dc42