darray and varray should only be keywords when followed by "["
Summary: `darray` and `varray` in an expression position are only interpreted as words with a special meaning when followed by `[`. This means that `darray` and `varray` can be used as function names. This should allow a definition of these functions that, under the hood, uses an `array` cast in order to take a `mixed` type object and return a `darray` or a `varray` with the required type parameters.

Reviewed By: KendallHopkins

Differential Revision: D4950084

fbshipit-source-id: c238ac0bb58e1ab8ed27f82a7865633c0b267ab9