Add T[Opt][Unc]ArrKey to hhbbc's type system
Summary:
HHBBC does not support arbitrary type-unions, but TInt | TStr comes up in a
number of contexts, so it makes sense to be able to represent that. Add support
to represent the type in RepoAuthType all the way through to the JIT. Change a
few obvious places in HHBBC to use TArrKey, but it can probably be used more.

Reviewed By: markw65

Differential Revision: D4940134

fbshipit-source-id: 4d0c1cbfcaf2c6e920c9fcd9bb2e47f7a95b0a85