Hack codegen: top-level const, correct ordering of DefCls and DefTypeAlias
Summary:
We now support `const` at top-level. Also the correct ordering of the instructions for top-level entities such as type aliases and
classes.

Also fix a couple of minor issues: ordering of `public static final`, and not picking up `declvars` in exception handlers.

And finally: added support for `GLOBALS[$var]` as an array base.

Reviewed By: oulgen, hubyrod

Differential Revision: D4946330

fbshipit-source-id: c4dfc3cd382a86c0be471a81a13da17f17bcc6af