Give empty namespace body an explicit token tag
Summary:
There are two valid namespace declarations:
```
namespace Foo {}
namespace Bar;
```
Since we want Full Fidelity parsing, we need to keep all the tokens somewhere in the resulting tree. Suppose the braced case could take a semicolon, i.e.
```
namespace Foo {};
```
we could just mark the body in the `Bar` case to be `Missing`. Now, we can't. This diff introduces the explicit tag `NamespaceEmptyBody` to have a more descriptive alternative than `Token`, but still have a hook to hang the semicolon on.

Reviewed By: KendallHopkins

Differential Revision: D4860378

fbshipit-source-id: f9781af4d378d1d73ff8e35276dad5039f4bfde8