Set a max amount of time per-request allowed in tc
Summary:
Our application and deployment process doesn't lend itself well to
a warmup script, and especially in our developer/test environments
the jit costs are expensive. This sets a per-request jit budget
that amortizes the jit cost across requests.

Tested with the following config flags:

	hhvm.jit_max_request_translation_time = 50000
	hhvm.jit_timer = 1

Background:
We have a large and broad PHP application that we deploy to dozens of times per-day. Because of a variety of deployment constraints that are difficult for us to change in the near-term, it's difficult for us to run a warm-up script before traffic is served.

Because of this, we observe high tail-latencies on requests after a deployment. Because we have a vary large pool of hosts in the fleet that we serve round-robin and have some infrequently accessed pages, some users end up hitting the jit penalty repeatedly.

Desired change:
We'd like to set a per-request budget for the amount of time to spend in the jit, to amortize that cost across requests.  We'd like to tradeoff a longer time getting code to be fully JITted for a smoother response-time upper bound post-deploy.  We took a stab at adding the limit and our running it on our dev environments, and would like feedback.

Here's what we are seeing after restart for 100 requests, time-to-first-byte vs the size of the tc cache reported in the admin port:

![max-request-time-stock](https://cloud.githubusercontent.com/assets/43085/25233870/28bf43ea-2595-11e7-952b-956071986bc8.png)

With the translation limit set to 100ms:

![max-request-time-with-change](https://cloud.githubusercontent.com/assets/43085/25233878/2f56dc54-2595-11e7-8e2a-89467d128592.png)

Thanks, and we'd love feedback on the idea and implementation!
Closes https://github.com/facebook/hhvm/pull/7796

Differential Revision: D4923007

Pulled By: mofarrell

fbshipit-source-id: 675e9e580b4cbdb685997328ee2176954e9a8e45