Use use-counts in fold-imms
Summary:
I noticed we didn't fold byte-sized immediates into stores in
prologues, and found a special check to prevent it.

This must be because prologues store zero to the type bytes of all the
locals, and so its better to use a register (assuming there's more
than one local).

Instead, we can count the uses, and fold when there's exactly
one.

Reviewed By: edwinsmith

Differential Revision: D4954038

fbshipit-source-id: 68dfe75afbbad1fc738769dccc23664edceb8f37