Create "suspend" operator for coroutine prototype
Summary:
I have added a unary prefix operator "suspend" to the full-fidelity parser; it has the same (low) precedence as "await".

This syntactic choice is provisional, and chosen to facilitate rapid testing of the prototype. It may not be the final choice; we still have a lot of experimentation and language design work to do.

I've added it only to the full-fidelity parser, so that the prototype coroutine lowering pass can remove it.

The semantics of the operator will be to introduce a suspension point in a coroutine when calling another coroutine:

    $x = suspend mycoroutine(123);

will be desugared into something like:

    $closure->nextLabel = 1;
    $data = mycoroutine(123, $closure);
    if ($data->is_suspended()) return $data();
    // If we didn't suspend then we keep going. If we did suspend,
    // then we'll resume and $data or $exception will be set to the
    // appropriate values depending on what happened in the coroutine
    // after it completes.
    label_1:
    if ($exception != null) throw $exception;
    $closure->x = $data->get_value();

Reviewed By: michaeltingley

Differential Revision: D4958680

fbshipit-source-id: 8ef4c8b3f1a418b6ef242e6df03bf910b34cfe71