jit: Removal of vasm instructions subb and subbi
Summary:
I wanted to verify the correctness of aarch64's emit function for subb and subbi, when I realised that both instructions are not emitted at all. With this PR I propose the elimination of them. Note that subb was not even implemented for x86_64 and ppc64.

The all set showed no regressions on ARM64.

mxw please review.
Closes https://github.com/facebook/hhvm/pull/7795

Differential Revision: D4949258

Pulled By: mxw

fbshipit-source-id: 56878fea90c42403f0c6fee4f09f2d743a5ec705