Assume name in identifier position of alias declaration
Summary:
For parsing `.hhi` files - and possibly other content describing the standard library / language components - the parser must allow the use of what are normally considered keywords in the identifier position. For example, `classname.hhi` defines
```
newtype classname<+T> as typename<T> = typename<T>;
```
In this case, `classname` is actually an identifier, not a keyword.

Reviewed By: ericlippert

Differential Revision: D4826736

fbshipit-source-id: cf8b5676bf36983eec3d0b5c59082866f203d141