Allow omission of visibility specifiers
Summary:
PHP does not require class members to be explicitly specified as being `public`, `protected` or `private`, even though Hack requires it. This means that
```
class Foo {
  async function bar() {
    return 42;
  }
}
```
is acceptable. This diff makes `Async` a lookahead for a class member.

The use, in the same test that surfaced this missing feature, of `async static` instead of `static async` is not compliant with Hack's syntax. Since `async` is only a modifier in Hack and not PHP, this diff corrects the use site.

Reviewed By: hubyrod

Differential Revision: D4859667

fbshipit-source-id: 9d35e3a91e669c864735a1a8754fc450db1858e7