Linting in lowering and Pos off-by-one bug in lowering
Summary:
D4716671 is blocked by test failures. Cutting over from `Parser_hack` to using the full fidelity parser skipped uppercase vs. lowercase linting for literal expressions. This is because that part of the linter is implementined //in the parser//.

Now, the lowerer includes this linter check and it fixes an off-by-one error in the position computation (that seems wrong, but should disappear when we replace the somewhat obscure `Pos.t` with something faster/more consistent).

Reviewed By: hubyrod

Differential Revision: D4985665

fbshipit-source-id: c71b2eb563e83c2c99fde2e9dabe03664d632074