Adjust cache location for PC relative branches
AArch64 supports up to +/-128MB PC relative branches with 26-bit
immediates in the B and BL instructions, so move aHot closer to
the aCold cache.

No new failures in the 'all' unit test suite, nor in the oss-
perfomrance benchmarking suite. When combined with relocation, it
produces nice gains in oss.