Two small fixes for XHP children expression parsing in full-fidelity-parser
Summary:
I made two small mistakes in the FFP for parsing XHP children expressions, which are here corrected.

First, there may be only a single trailing unary postfix operator in an expression. That is "foo*" is good, but "foo*+" is not legal. The FFP was accepting any number of trailing operators.

Second, I parsed the sole binary operator (|) as right-associative, but the original HHVM parser parses it as left-associative.

Both problems are fixed.

This problem was blocking oulgen's work on code generation.

Reviewed By: oulgen

Differential Revision: D4989178

fbshipit-source-id: 094bcc2e714acd0a09cbd5bc82807e121a654878