Fix race in smash{Jmp,Jcc,Call} and retranslate() and reintroduce D4704952, D4724004, and D4762274
Summary:
Reverts D4892661. The underlying bug was that to make smash{Call,Jmp,Jcc} atomic
DataBlock was performing an eight byte write (call and jmp are 5 bytes, and jcc
is 6). This was problematic when the codeLock was no longer required for these
instructions, as the read-modify-write of the extra 2-3 bytes created a data
race when the smashed instruction fell at the end of a translation or adjacent
to another smash instruction.

Additionally D4704952 removed a lock in retranslate() that was used to avoid
creating duplicate retranslations. We can achieve the same safety by checking
the number of translations before and after acquiring the WriteLease and
comparing it with the number afterwards.

To make smashCall atomic it was all necessary to fix t7889486 by iterating the
ProfTransRecs and updating any mainCaller or guardCaller addresses found after
liveRelocation.

Reviewed By: markw65

Differential Revision: D4961657

fbshipit-source-id: 7a858ceaf79ceb5de344ce86f66a25d68404c425