Fixed erroneous default instantiation of Relative_path.t in lowerer
Summary:
The `Parser_heap` and `File_heap` use `Relative_pos.t` to store files. The lowerer gets a `Relative_pos.t`, takes the path out and then re-instantiates a `Dummy` instance with said path when it needs to store the `Relative_pos.t`. This is obviously a bug.

When stacking on top of D4716671, it prunes the `Decl_not_found` errors.

Reviewed By: hubyrod

Differential Revision: D4993557

fbshipit-source-id: 40ac09093dee5b4e688a502657bca2a9dfbf0ebe