Hack codegen: adata
Summary:
Big simplification to the way we generate inline literals, in `adata` and elsewhere. Fixed issue with non-integer/string keys in map-like arrays. Also static vs non-static collections are not dependent on a separate `is_literal` predicate, but use the `Typed_value` functionality to
attempt to transform an expression to a typed value.

Reviewed By: hubyrod

Differential Revision: D4993381

fbshipit-source-id: 13e98db509617d1bbab7d899f2e44ca5b4ecf121