Add support for int-like string array key HackArrCompatNotices
Summary:
Add HackArrCompatNotices support for using an int-like string as an array
key. This happens in a number of builtins (and we don't care about those), so
add suppressions as needed.

HPHPc and HHBBC shouldn't treat arrays with known int-like string array keys as
scalars, as they will trigger the notice when constructed.

With this change, a few IR ops which couldn't re-enter before, can now. To avoid
regressions with this option disabled, special case those ops when calculating
memory effects and constructing catch traces.

The code which checks for int-like string array keys was very hot, so simply
doing the runtime option check there is a regression. Instead templatize the
function, and propagate the boolean all the way up the callers to the generated
code, which can emit a call to the right template instantiation depending on the
runtime option. This is ugly and invasive, but it avoids regressions.

Reviewed By: mxw

Differential Revision: D4830984

fbshipit-source-id: 64b76793e0cf1f7bba3f51ff5ee4d3348b6dc0a5