A Mock watchman module that is injectable
Summary:
The Informant is difficult to test. I want to avoid making it a functor
(and proliferating functorization throughout the entire codebase which
will get us to "constructor hell"). Which is what the injector is
there to solve. Here I make an injected
Watchman module that is Mocked. See in-line comments for more info.

Differential Revision: D4933141

fbshipit-source-id: 6c75dc6e409970ac2a67a17c3cc3cec4c7d812bd