Tidy up process-wide WatchmanClient cache in ext_watchman
Summary:
When running Facts DB on Sandcastle runs I've noticed a small number of new subscription attempts get stuck. Examining core dumps in gdb revealed there were oustanding promises in `s_connectPromises` but no `WatchmanClient` instances to fulfil these in `s_allClients`. I was not able to figure out how this situation was triggered but the whole system of connection mangement was too complicated.

In this diff we re-write the ext_watchman connection management code to be simpler and clearer.

Reviewed By: paulbiss

Differential Revision: D4816886

fbshipit-source-id: 2941d1a3644e878c600fe8a7547efb24a29dbd43