Introduce the state_machine_data type and use it to relay information about the function's body to the closure generator
Summary:
Prior to this diff, there was ambiguity between what the `coroutine_state_machine_generator`, the `coroutine_closure_generator`, and the `coroutine_method_lowerer` were extracting. The `coroutine_state_machine_generator` and `coroutine_closure_generator` were both dealing with locals and params. Since the `coroutine_state_machine_generator` is most directly dealing with the coroutine's body, it makes sense for it to extract information from the function, and pass it along in order for the other functions to use this information to generate their code.

This diff introduces the `state_machine_data` type, which will allow the `coroutine_state_machine_generator` to include additional information that it has extracted about the function. The later modules in this flow can then use that extracted data, rather than recomputing it themselves, in order to generate code.

Reviewed By: ericlippert

Differential Revision: D4994950

fbshipit-source-id: c1fc4e3d97da5a655d972be6d1284835a6a3b87f